+++ lib/langtool_web/controllers/dashboard_controller.ex
@@ -1,6 +1,8 @@
 defmodule LangtoolWeb.DashboardController do
   use LangtoolWeb, :controller
 
+  plug :check_auth when action in [:index]
+
   def index(conn, _) do
     render conn, "index.html"
   end