+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - signout process
 - signin process
 - presenter for validation errors
+- dashboard page with auth check for access
 
 ## [0.1.0] - 2019-01-01
 ### Added